+++ src/impls/maps/hash_map.rs
@@ -6,7 +6,7 @@ use fluent_result::IntoResult;
 use size_guess::SizeGuess;
 use tap::Pipe;
 
-use crate::{KeyCollision, TryExtend, TryFromIterator};
+use crate::{FoldMut, KeyCollision, TryExtend, TryFromIterator};
 
 /// Converts an iterator of key-value pairs into a [`HashMap`], failing if a key would collide.
 #[allow(clippy::implicit_hasher)]
@@ -28,10 +28,9 @@ impl<K: Eq + Hash, V> TryFromIterator<(K, V)> for HashMap<K, V> {
         let mut iter = into_iter.into_iter();
         let size_guess = iter.size_guess();
 
-        iter.try_fold(HashMap::with_capacity(size_guess), |mut map, (k, v)| match map.entry(k) {
+        iter.try_fold_mut(HashMap::with_capacity(size_guess), |map, (k, v)| match map.entry(k) {
             Entry::Occupied(entry) => entry.remove_entry().0.pipe(KeyCollision::new).into_err(),
-            #[rustfmt::skip]
-            Entry::Vacant(entry) => { entry.insert(v); Ok(map) }
+            Entry::Vacant(entry) => Ok(_ = entry.insert(v)),
         })
     }
 }
@@ -53,12 +52,11 @@ impl<K: Eq + Hash, V, S: BuildHasher> TryExtend<(K, V)> for HashMap<K, V, S> {
         let mut iter = iter.into_iter();
         let size_guess = iter.size_guess();
 
-        iter.try_fold(HashMap::with_capacity(size_guess), |mut map, (key, value)| match self.contains_key(&key) {
+        iter.try_fold_mut(HashMap::with_capacity(size_guess), |map, (key, value)| match self.contains_key(&key) {
             true => KeyCollision::new(key).into_err(),
             false => match map.entry(key) {
                 Entry::Occupied(entry) => entry.remove_entry().0.pipe(KeyCollision::new).into_err(),
-                #[rustfmt::skip]
-                Entry::Vacant(entry) => { entry.insert(value); Ok(map) }
+                Entry::Vacant(entry) => Ok(_ = entry.insert(value)),
             },
         })
         .map(|insert_map| self.extend(insert_map))
@@ -66,7 +64,7 @@ impl<K: Eq + Hash, V, S: BuildHasher> TryExtend<(K, V)> for HashMap<K, V, S> {
 
     /// Appends an iterator of key-value pairs to the map, failing if a key would collide.
     ///
-    /// This implementation provides a weak error guarantee. If the method returns an error, the
+    /// This implementation provides a basic error guarantee. If the method returns an error, the
     /// map may be modified. However, it will still be in a valid state, and the specific
     /// collision that caused the error will not take effect.
     ///