+++ src/impls/sets/hash_set.rs
@@ -4,7 +4,7 @@ use std::hash::{BuildHasher, Hash};
 use fluent_result::IntoResult;
 use size_guess::SizeGuess;
 
-use crate::{TryExtend, TryFromIterator, ValueCollision};
+use crate::{FoldMut, TryExtend, TryFromIterator, ValueCollision};
 
 /// Converts an iterator of values into a [`HashSet`], failing if a key would collide.
 #[allow(clippy::implicit_hasher)]
@@ -26,13 +26,9 @@ impl<T: Eq + Hash> TryFromIterator<T> for HashSet<T> {
         let mut iter = into_iter.into_iter();
         let size_guess = iter.size_guess();
 
-        iter.try_fold(HashSet::with_capacity(size_guess), |mut set, value| {
-            match set.contains(&value) {
-                true => return Err(ValueCollision::new(value)),
-                false => set.insert(value),
-            };
-
-            Ok(set)
+        iter.try_fold_mut(HashSet::with_capacity(size_guess), |set, value| match set.contains(&value) {
+            true => Err(ValueCollision::new(value)),
+            false => Ok(_ = set.insert(value)),
         })
     }
 }
@@ -54,12 +50,11 @@ impl<T: Eq + Hash, S: BuildHasher> TryExtend<T> for HashSet<T, S> {
         let mut iter = iter.into_iter();
         let size_guess = iter.size_guess();
 
-        iter.try_fold(HashSet::with_capacity(size_guess), |mut set, value| match self.contains(&value) {
+        iter.try_fold_mut(HashSet::with_capacity(size_guess), |set, value| match self.contains(&value) {
             true => ValueCollision::new(value).into_err(),
             false => match set.contains(&value) {
                 true => ValueCollision::new(value).into_err(),
-                #[rustfmt::skip]
-                false => { _ = set.insert(value); Ok(set) }
+                false => Ok(_ = set.insert(value)),
             },
         })
         .map(|set| self.extend(set))
@@ -67,7 +62,7 @@ impl<T: Eq + Hash, S: BuildHasher> TryExtend<T> for HashSet<T, S> {
 
     /// Appends an iterator of values to the set, failing if a value would collide.
     ///
-    /// This implementation provides a weak error guarantee. If the method returns an error, the
+    /// This implementation provides a basic error guarantee. If the method returns an error, the
     /// set may be modified. However, it will still be in a valid state, and the specific
     /// collision that caused the error will not take effect.
     ///