+++ src/traits/mod.rs
@@ -1,8 +1,10 @@
+mod fold_mut;
 mod try_collect_ex;
 mod try_extend;
 mod try_from_iterator;
 mod try_unzip;
 
+pub use fold_mut::FoldMut;
 pub use try_collect_ex::TryCollectEx;
 pub use try_extend::TryExtend;
 pub use try_from_iterator::TryFromIterator;