+++ src/traits/try_collect_ex.rs
@@ -1,33 +1,38 @@
+use include_doc::function_body;
+
+#[cfg(doc)]
+use std::collections::HashMap;
+
 use crate::TryFromIterator;
 
 /// Extends [Iterator] with a failable collect method.
-pub trait TryCollectEx {
-    /// The item type of the iterator.
-    type Item;
-
-    /// Collects the iterator into a container, returning an error if construcing the container fails.
+///
+/// This trait lets you have an iterator return any collection that can be created via
+/// [`TryFromIterator`], similar to [`Iterator::collect`] and [`FromIterator::from_iter`],
+/// but with the ability to return a implementation specific error if the creation of the contaienr
+/// fails some invariant.
+pub trait TryCollectEx: Iterator {
+    /// Tries to collects the iterator into a container, returning an error if construcing the
+    /// container fails.
     ///
-    /// This is a wrapper for [`TryFromIterator::try_from_iter`].
+    /// Exact behavior of this method depends on the container implementation, but generally it
+    /// should be expected to short-circuit on the first error.
     ///
-    /// This is useful if you want to collect an iterator into a container that may fail to be constructed.
+    /// On success, this method should behave similarly to [`Iterator::collect`], except returning
+    /// a [`Result`].
     ///
-    /// Ideally this would be called `try_collect` but there is a method with that name in nightly.
+    /// Note: Ideally this would be called `try_collect` but there is a method with that name in nightly.
     ///
     /// # Errors
     ///
-    /// Returns an error if the container fails to be constructed.
+    /// Returns a [`TryFromIterator::Error`] if the container fails to be constructed.
     ///
     /// # Example
     ///
+    /// Collecting into a [`HashMap`] that fails if a key would collide.
+    ///
     /// ```rust
-    /// use std::collections::HashMap;
-    /// use collect_failable::{TryCollectEx, TryFromIterator, KeyCollision};
-    ///
-    /// let result = [(1, 2), (2, 3)].into_iter().try_collect_ex::<HashMap<_, _>>();
-    /// assert_eq!(result, Ok(HashMap::from([(1, 2), (2, 3)])));
-    ///
-    /// let result = [(1, 2), (1, 3)].into_iter().try_collect_ex::<HashMap<_, _>>();
-    /// assert_eq!(result, Err(KeyCollision { key: 1 }));
+    #[doc = function_body!("tests/try_collect_ex.rs", try_collect_ex_collision_example, [])]
     /// ```
     fn try_collect_ex<C>(self) -> Result<C, C::Error>
     where
@@ -40,8 +45,6 @@ impl<I, T> TryCollectEx for I
 where
     I: Iterator<Item = T>,
 {
-    type Item = T;
-
     fn try_collect_ex<C>(self) -> Result<C, C::Error>
     where
         C: TryFromIterator<Self::Item>,