+++ src/traits/try_unzip.rs
@@ -1,26 +1,36 @@
+use include_doc::function_body;
+
 use crate::{OneOf2, TryExtend};
 
-/// Extends [Iterator] with a failable unzip method.
+/// Extends [`Iterator`] with a failable unzip method.
+///
+/// This is similar to [`Iterator::unzip`], but allows for failable construction. The created
+/// containers may be of different types, but both must implement [`Default`] and [`TryExtend`].
 pub trait TryUnzip {
     /// Tries to unzip the iterator into two collections.
     ///
-    /// This is a failable version of [`Iterator::unzip`].
+    /// Both containers are extended, element by element, in parallel.
     ///
     /// # Errors
     ///
-    /// Returns an error if any of the underlying collections fail to extend.
+    /// Returns a [`TryExtend::Error`] wrapped in a [`OneOf2`] if either of the underlying
+    /// collections fail to extend.
     ///
     /// # Examples
     ///
+    /// Different types of containers can be unzipped into.
+    ///
+    /// ```rust
+    #[doc = function_body!("tests/try_unzip.rs", try_unzip_different_containers_example, [])]
     /// ```
-    /// use collect_failable::{TryUnzip, OneOf2, ValueCollision};
-    /// use std::collections::HashSet;
     ///
-    /// let data = vec![(1, 2), (2, 3)];
-    /// let (a, b): (HashSet<i32>, HashSet<i32>) = data.into_iter().try_unzip().unwrap();
+    /// ## Multiple Errors
     ///
-    /// assert_eq!(a, HashSet::from([1, 2]));
-    /// assert_eq!(b, HashSet::from([2, 3]));
+    /// In the case of multiple possible failures, the error from extending `FromA` is checked and
+    /// returned first.
+    ///
+    /// ```rust
+    #[doc = function_body!("tests/try_unzip.rs", try_unzip_collision_example, [])]
     /// ```
     fn try_unzip<A, B, FromA, FromB>(self) -> Result<(FromA, FromB), OneOf2<FromA::Error, FromB::Error>>
     where
@@ -29,6 +39,7 @@ pub trait TryUnzip {
         Self: Iterator<Item = (A, B)>;
 }
 
+/// Implemententation for any [`Iterator`].
 impl<I> TryUnzip for I
 where
     I: Iterator,